requirements.txt
langchain==0.1.0
langchain-dashscope==0.1.0
gradio==4.0.0
python-dotenv==1.0.0
chromadb>=0.4.24
pydantic>=2.0.0